add context for todos
