+++ src/types/Filter.ts
@@ -1,1 +1,5 @@
-export type Filter = 'all' | 'active' | 'completed';
+export enum Filter {
+  ALL = 'all',
+  ACTIVE = 'active',
+  COMPLETED = 'completed',
+}
